pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "takebridge-orchestrator"
version = "0.1.0"
description = "Agent orchestration service for UI automation with worker, grounding, and code agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "TakeBridge" }
]
dependencies = [
  "fastapi==0.120.0",
  "uvicorn[standard]==0.38.0",
  "openai==2.6.1",
  "httpx==0.28.1",
  "requests==2.32.5",
  "python-dotenv==1.1.1",
  "numpy==2.3.4",
  "pillow==12.0.0",
  "pytesseract==0.3.13",
  "opencv-python==4.11.0.86",
  "pyautogui==0.9.54",
  "typing-extensions==4.15.0",
  "annotated-doc==0.0.3",
  "annotated-types==0.7.0",
  "anyio==4.11.0",
  "certifi==2025.10.5",
  "charset-normalizer==3.4.4",
  "click==8.3.0",
  "distro==1.9.0",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httptools==0.7.1",
  "idna==3.11",
  "jiter==0.11.1",
  "mouseinfo==0.1.3",
  "packaging==25.0",
  "pydantic==2.12.3",
  "pydantic-core==2.41.4",
  "pygetwindow==0.0.9",
  "pymsgbox==2.0.1",
  "pyobjc-core==12.0",
  "pyobjc-framework-cocoa==12.0",
  "pyobjc-framework-quartz==12.0",
  "pyperclip==1.11.0",
  "pyrect==0.2.0",
  "pyscreeze==1.0.1",
  "pytweening==1.2.0",
  "pyyaml==6.0.3",
  "rubicon-objc==0.5.2",
  "sniffio==1.3.1",
  "starlette==0.48.0",
  "tqdm==4.67.1",
  "typing-inspection==0.4.2",
  "urllib3==2.5.0",
  "uvloop==0.22.1",
  "watchfiles==1.1.1",
  "websockets==15.0.1",
  "mcp>=1.20.0",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg2-binary>=2.9.11",
  "tiktoken>=0.12.0",
  "sentence-transformers>=2.2.0",
  "python-jose[cryptography]>=3.3.0",
  "boto3>=1.34.0",
  "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.3.0",
  "ruff>=0.4.0",
  "pytest>=7.4.0",
  "mypy>=1.8.0",
]

[project.scripts]
takebridge-server = "server.api.server:app"

[tool.hatch.build.targets.wheel]
packages = ["computer_use_agent", "mcp_agent", "server", "shared"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I"]
